v3.0.2
Removed SqlCallContextAttribute: ISqlCallContext is enough.
Changed the namespace of CK.SqlServer.Setup.ISqlCallContext: it is now in CK.SqlServer: it is easier to use.
Use OnError on monitor instead of Catch & CatchCounter.
Added warning when auto implementation has not received a call to DynamicItemInitialize.
Added a check (exception) if the database can not be created (ie. is back to master!).
Accepts [Replace] and [Override] prefix for resource names. This is currently only an option that has absolutely no impact but could be enforced once.