Add RawCrisReceiver to magic type registration.
CK.DB.Tests.NUnit should be CK.Testing.SqlServerEngine.NUnit.
CK.Testing.SqlServerEngine.NUnit SHOULD NOT bring the
CK.Cris.Executor.Engine (and the 2  magically available types).

But all this works like a charm and, one day, the Sql layer will be
heavily refactored, so we keep this as-is for the moment.